refactor integration test and add test
