+++ cmd/integration-test-server/main.go
@@ -14,8 +14,6 @@ import (
 	"math"
 	"net/http"
 	"os"
-	"regexp"
-	"strings"
 	"time"
 
 	"github.com/crate-crypto/go-proto-danksharding-crypto/eth"
@@ -25,6 +23,7 @@ import (
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/ethereum/go-ethereum/rpc"
 	"github.com/ethstorage/go-ethstorage/cmd/es-utils/utils"
+	itutils "github.com/ethstorage/go-ethstorage/cmd/integration-test-server/utils"
 	es "github.com/ethstorage/go-ethstorage/ethstorage"
 	"github.com/ethstorage/go-ethstorage/ethstorage/node"
 	prv "github.com/ethstorage/go-ethstorage/ethstorage/prover"
@@ -178,7 +177,10 @@ func checkFinalState(state *node.NodeState) {
 			addErrorMessage("At lease one block should be mined successfully during the test.")
 		}
 		if shardState.SubmissionState.Failed > 0 {
-			knownFailureCount := checkKnownFailure()
+			knownFailureCount, err := itutils.CheckKnowFailure(logFile)
+			if err != nil {
+				addErrorMessage(err.Error())
+			}
 			if knownFailureCount > 0 {
 				log.Warn("Check known failure result", "failure", shardState.SubmissionState.Failed, "known failure", knownFailureCount)
 			}
@@ -286,203 +288,6 @@ func downloadBlobFromRPC(client *rpc.Client, kvIndex uint64, hash common.Hash) (
 	}
 
 	return result, nil
-}
-
-func checkKnownFailure() int {
-	count0, err := checkDiffNotMatchError()
-	if err != nil {
-		addErrorMessage(fmt.Sprintf("checkDiffNotMatchError fail: err, %s", err.Error()))
-	}
-	count1, err := checkInvalidSamplesError()
-	if err != nil {
-		addErrorMessage(fmt.Sprintf("checkInvalidSamplesError fail: err, %s", err.Error()))
-	}
-
-	return count0 + count1
-}
-
-func checkDiffNotMatchError() (int, error) {
-	// Description: "Mining info retrieved" -> "Failed to submit mined result"; shard and block equal which difficulty is not the same
-	// Sample:
-	// lvl=info msg="Mining info retrieved"                 shard=1 block=6,906,682 difficulty=13,006,115 lastMineTime=1,729,375,716 proofsSubmitted=2
-	// lvl=eror msg="Failed to submit mined result"         shard=1 block=6,906,682 difficulty=1,729,376,532 error="failed to estimate gas: execution reverted: StorageContract: diff not match"
-
-	file, err := os.OpenFile(logFile, os.O_RDONLY, 0755)
-	if err != nil {
-		return 0, err
-	}
-	defer file.Close()
-
-	fileScanner := bufio.NewScanner(file)
-	fileScanner.Split(bufio.ScanLines)
-
-	count := 0
-	difficultyMap := make(map[string]string)
-	for fileScanner.Scan() {
-		logText := fileScanner.Text()
-		if strings.Contains(logText, "Mining info retrieved") {
-			block, diff, err := fetchBlockAndDifficulty(logText)
-			if err != nil {
-				log.Error("fetchBlockAndDifficulty error", "log", logText, "error", err.Error())
-				continue
-			}
-			difficultyMap[block] = diff
-		} else if regexp.MustCompile(`Failed to submit mined result[\s\S]+diff not match`).MatchString(logText) {
-			block, diff, err := fetchBlockAndDifficulty(logText)
-			if err != nil {
-				log.Error("fetchBlockAndDifficulty error", "log", logText, "error", err.Error())
-				continue
-			}
-			if originalDiff, ok := difficultyMap[block]; ok && strings.Compare(diff, originalDiff) != 0 {
-				log.Warn("Known diff not match error", "block", block, "original difficulty", originalDiff, "latest difficulty", diff)
-				count++
-				continue
-			}
-		}
-	}
-
-	return count, nil
-}
-
-func checkInvalidSamplesError() (int, error) {
-	// description: 1. sample empty blob k, 2. download blob k, 3. submit mined result fail
-	// Sample:
-	// lvl=info msg="Get data hash"                         kvIndex=11742 hash=0x0000000000000000000000000000000000000000000000000000000000000000
-	// lvl=info msg="Downloaded and encoded"                blockNumber=4,225,672 kvIdx=11742
-	// lvl=info msg="Got storage proof"                     shard=1 block=6,906,682 kvIdx="[14613 11742]" sampleIdxsInKv="[1691 1859]"
-	// lvl=info msg="Submit mined result done"              shard=1 block=6,906,682 nonce=739,669   txSigner=0x8be2c9379eb69877F25aBa61a853eC4FCb0b273a hash=0x4f212aa8f5b8867b6478f4cde9dc09609590ff4fc14997f7048ff32f2c96af0c
-	// lvl=eror msg="Failed to submit mined result"         shard=1 block=6,906,682 error="failed to estimate gas: execution reverted: EthStorageContract2: invalid samples"
-
-	file, err := os.OpenFile(logFile, os.O_RDONLY, 0755)
-	if err != nil {
-		return 0, err
-	}
-	defer file.Close()
-
-	fileScanner := bufio.NewScanner(file)
-	fileScanner.Split(bufio.ScanLines)
-
-	count := 0
-	minedEmptyKVs := make(map[string]string)
-	legacyKVs := make(map[string]string)
-	for fileScanner.Scan() {
-		logText := fileScanner.Text()
-		if regexp.MustCompile(`Get data hash[\s\S]+hash=0x0000000000000000000000000000000000000000000000000000000000000000`).MatchString(logText) {
-			kvIdx := extractWithName(logText, `kvIndex=(?P<kvIdx>[\d]+)`, "kvIdx")
-			if kvIdx != "" {
-				minedEmptyKVs[kvIdx] = ""
-			}
-		} else if strings.Contains(logText, `Downloaded and encoded`) {
-			kvIdx := extractWithName(logText, `kvIdx=(?P<kvIdx>[\d]+)`, "kvIdx")
-			if kvIdx == "" {
-				continue
-			}
-			block, ok := minedEmptyKVs[kvIdx]
-			if !ok {
-				continue
-			}
-			legacyKVs[kvIdx] = block
-		} else if strings.Contains(logText, `Got storage proof`) {
-			block, kvIdxes, err := fetchMinedBlockAndKVIdx(logText)
-			if err != nil {
-				log.Error("checkInvalidSamplesError error", "log", logText, "error", err.Error())
-			}
-			for kvIdx := range minedEmptyKVs {
-				if !strings.Contains(kvIdxes, kvIdx) {
-					continue
-				}
-				minedEmptyKVs[kvIdx] = block
-				if _, ok := legacyKVs[kvIdx]; ok {
-					legacyKVs[kvIdx] = block
-				}
-			}
-		} else if regexp.MustCompile(`Failed to submit mined result[\s\S]+invalid samples`).MatchString(logText) {
-			for kvIdx, block := range legacyKVs {
-				if block != "" && strings.Contains(logText, block) {
-					log.Warn("Known error", "block", block, "kvIdx", kvIdx, "error", "invalid samples")
-					delete(legacyKVs, kvIdx)
-					count++
-					continue
-				}
-				delete(minedEmptyKVs, kvIdx)
-			}
-		} else if strings.Contains(logText, "Submit mined result done") {
-			for kvIdx, block := range minedEmptyKVs {
-				if block != "" && strings.Contains(logText, block) {
-					delete(minedEmptyKVs, kvIdx)
-				}
-				continue
-			}
-		}
-	}
-
-	return count, nil
-}
-
-func fetchMinedBlockAndKVIdx(text string) (block string, kvIdx string, err error) {
-	patten := `block=(?P<block>[\d{1,3}(,\d{3})*]+)[\s]+kvIdx=\"\[(?P<kvIdx>\d+ \d+)\]\"`
-
-	results := extract(text, patten)
-	if len(results) == 2 {
-		for name, value := range results {
-			if strings.Compare(name, "block") == 0 {
-				block = value
-			} else if strings.Compare(name, "kvIdx") == 0 {
-				kvIdx = value
-			}
-		}
-	}
-
-	if block == "" || kvIdx == "" {
-		err = errors.New(fmt.Sprintf("extract mined block and nonce fail, %s, error: wrong log format", patten))
-	}
-
-	return
-}
-
-func fetchBlockAndDifficulty(text string) (string, string, error) {
-	patten := `shard=(?P<shard>[\d])[\s]+block=(?P<block>[\d{1,3}(,\d{3})*]+)[\s]+difficulty=(?P<difficulty>[\d{1,3}(,\d{3})*]+)`
-	shard, block, diff := "", "", ""
-
-	results := extract(text, patten)
-	if len(results) == 3 {
-		for name, value := range results {
-			if strings.Compare(name, "shard") == 0 {
-				shard = value
-			} else if strings.Compare(name, "block") == 0 {
-				block = value
-			} else if strings.Compare(name, "difficulty") == 0 {
-				diff = value
-			}
-		}
-	}
-
-	if shard == "" || block == "" || diff == "" {
-		return "", "", errors.New(fmt.Sprintf("extract block and difficulty fail, patten: %s; error: wrong log format.", patten))
-	}
-
-	return fmt.Sprintf("%s-%s", shard, block), diff, nil
-}
-
-func extractWithName(text, patten, name string) string {
-	results := extract(text, patten)
-	return results[name]
-}
-
-func extract(text, patten string) map[string]string {
-	results := make(map[string]string)
-	re := regexp.MustCompile(patten)
-
-	matches := re.FindStringSubmatch(text)
-	if matches != nil {
-		for i, name := range re.SubexpNames() {
-			if i != 0 && name != "" {
-				results[name] = matches[i]
-			}
-		}
-	}
-
-	return results
 }
 
 func addErrorMessage(errMessage string) {
